requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==4.2.19
djangorestframework==3.15.2
exceptiongroup==1.2.2
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
pytest==8.3.4
pytest-django==4.10.0
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.12.2
tzdata==2025.1
